@@ -1,41 +1,66 @@
-# coding: utf-8
-import requests
-from requests.adapters import HTTPAdapter, DEFAULT_POOLSIZE
-from requests.exceptions import RequestException
-from requests.compat import urljoin
-from time import sleep
+"""Contains the horizon api abstraction"""
 
-from urllib3.exceptions import NewConnectionError
-from urllib3.util import Retry
+import json
+import asyncio
+from functools import wraps
+
+import aiohttp
+from yarl import URL
+
+from aiohttp_sse_client.client import EventSource as SSEClient
+
 from .version import __version__
 from .asset import Asset
 from .exceptions import HorizonError, HorizonRequestError
 
 import logging
+from typing import Union, Optional, AsyncGenerator
 
 logger = logging.getLogger(__name__)
 
-try:
-    from sseclient import SSEClient
-except ImportError:
-    SSEClient = None
 
-HORIZON_LIVE = "https://horizon.stellar.org"
-HORIZON_TEST = "https://horizon-testnet.stellar.org"
+HORIZON_LIVE = "https://horizon.kinfederation.com"
+HORIZON_TEST = "https://horizon-testnet.kininfrastructure.com"
 DEFAULT_REQUEST_TIMEOUT = 11  # two ledgers + 1 sec, let's retry faster and not wait 60 secs.
 DEFAULT_NUM_RETRIES = 3
 DEFAULT_BACKOFF_FACTOR = 0.5
-USER_AGENT = 'py-stellar-base-{}'.format(__version__)
+USER_AGENT = 'py-kin-base-{}'.format(__version__)
+HEADERS = {'User-Agent': USER_AGENT, 'Content-Type': 'application/x-www-form-urlencoded'}
+
+
+def _retry(func):
+    """Retry the function on specific errors"""
+    @wraps(func)
+    async def wrapper(*args, **kwargs):
+        self = args[0]
+        for i in range(self.num_retries + 1):
+            try:
+                return await func(*args, **kwargs)
+            except Exception as e:
+                if i == self.num_retries:
+                    raise
+                elif isinstance(e, (aiohttp.ClientConnectionError, aiohttp.ContentTypeError, asyncio.TimeoutError)):
+                    """
+                    We couldn't connect to horizon/ 
+                    The response was not a valid json (not an horizon response)/
+                    Timeout
+                    """
+                    pass
+                else:
+                    raise
+                logger.debug('# Retry {}: {}'.format(func.__name__, i))
+                await asyncio.sleep(self.backoff_factor * i)
+
+    return wrapper
 
 
 class Horizon(object):
     def __init__(self,
-                 horizon_uri=None,
-                 pool_size=DEFAULT_POOLSIZE,
-                 num_retries=DEFAULT_NUM_RETRIES,
-                 request_timeout=DEFAULT_REQUEST_TIMEOUT,
-                 backoff_factor=DEFAULT_BACKOFF_FACTOR,
-                 user_agent=USER_AGENT):
+                 horizon_uri: Optional[str] = None,
+                 pool_size: Optional[int] = None,
+                 num_retries: Optional[int] = DEFAULT_NUM_RETRIES,
+                 request_timeout: Optional[Union[int, None]] = DEFAULT_REQUEST_TIMEOUT,
+                 backoff_factor: Optional[float] = DEFAULT_BACKOFF_FACTOR):
         """The :class:`Horizon` object, which represents the interface for
         making requests to a Horizon server instance.
 
@@ -49,147 +74,179 @@ class Horizon(object):
         error or a valid response. Any other errors however are raised by this
         class.
 
-        :param str horizon_uri: The horizon base URL
-        :param int request_timeout: The timeout for all requests.
-        :param int pool_size: persistent connection to Horizon and connection pool
-        :param int num_retries: configurable request retry functionality
-        :param float backoff_factor: a backoff factor to apply between attempts after the second try
-        :param str user_agent: String representing the user-agent you want, such as "py-stellar-base"
+        :param horizon_uri: The horizon base URL
+        :param request_timeout: The timeout for all requests.
+        :param pool_size: persistent connection to Horizon and connection pool
+        :param num_retries: configurable request retry functionality
+        :param backoff_factor: a backoff factor to apply between attempts after the second try
 
         """
         if horizon_uri is None:
-            self.horizon_uri = HORIZON_TEST
+            self.horizon_uri = URL(HORIZON_TEST)
         else:
-            self.horizon_uri = horizon_uri
+            self.horizon_uri = URL(horizon_uri)
 
-        self.pool_size = pool_size
         self.num_retries = num_retries
-        self.request_timeout = request_timeout
         self.backoff_factor = backoff_factor
 
-        # adding 504 to the tuple of statuses to retry
-        self.status_forcelist = tuple(Retry.RETRY_AFTER_STATUS_CODES) + (504,)
-
-        # configure standard session
-
-        # configure retry handler
-        retry = Retry(
-            total=self.num_retries,
-            backoff_factor=self.backoff_factor,
-            redirect=0,
-            status_forcelist=self.status_forcelist)
-        # init transport adapter
-        adapter = HTTPAdapter(
-            pool_connections=self.pool_size,
-            pool_maxsize=self.pool_size,
-            max_retries=retry)
-
         # init session
-        session = requests.Session()
+        if pool_size is None:
+            connector = aiohttp.TCPConnector()
+        else:
+            connector = aiohttp.TCPConnector(limit=pool_size)
+        session = aiohttp.ClientSession(headers=HEADERS,
+                                        connector=connector,
+                                        timeout=aiohttp.ClientTimeout(total=request_timeout))
 
-        # set default headers
-        session.headers.update({'User-Agent': user_agent})
-
-        session.mount('http://', adapter)
-        session.mount('https://', adapter)
         self._session = session
+        self._sse_session = None
 
-        # configure SSE session (differs from our standard session)
+    async def _init_sse_session(self) -> None:
+        """Init the sse session """
+        if self._sse_session is None:
+            self._sse_session = aiohttp.ClientSession(headers={'User-Agent': USER_AGENT})  # No timeout, no special connector
+            # Other headers such as "Accept: text/event-stream" are added by thr SSEClient
 
-        sse_retry = Retry(
-            total=1000000, redirect=0, status_forcelist=self.status_forcelist)
-        sse_adapter = HTTPAdapter(
-            pool_connections=self.pool_size,
-            pool_maxsize=self.pool_size,
-            max_retries=sse_retry)
-        sse_session = requests.Session()
-        sse_session.headers.update({'User-Agent': user_agent})
-        sse_session.mount('http://', sse_adapter)
-        sse_session.mount('https://', sse_adapter)
-        self._sse_session = sse_session
+    async def __aenter__(self) -> 'Horizon':
+        return self
 
-    def submit(self, te):
+    async def __aexit__(self, exc_type, exc_val, exc_tb) -> None:
+        await self.close()
+
+    @_retry
+    async def submit(self, te: str) -> dict:
         """Submit the transaction using a pooled connection, and retry on failure.
 
         `POST /transactions
         <https://www.stellar.org/developers/horizon/reference/endpoints/transactions-create.html>`_
 
         Uses form-encoded data to send over to Horizon.
-
+        :param te: The transaction envelop encoded in base64
         :return: The JSON response indicating the success/failure of the
             submitted transaction.
-        :rtype: dict
 
         """
         params = {'tx': te}
-        url = urljoin(self.horizon_uri, 'transactions/')
+        abs_url = self.horizon_uri.join(URL('/transactions'))
+        try:
+            reply = await self._post(abs_url, params)
+        except (aiohttp.ClientConnectionError, aiohttp.ContentTypeError) as e:
+                raise HorizonRequestError(e)
 
-        # POST is not included in Retry's method_whitelist for a good reason.
-        # our custom retry mechanism follows
-        reply = None
-        retry_count = self.num_retries
-        while True:
-            try:
-                reply = self._session.post(
-                    url, data=params, timeout=self.request_timeout)
-                return check_horizon_reply(reply.json())
-            except (RequestException, NewConnectionError, ValueError) as e:
-                if reply is not None:
-                    msg = 'Horizon submit exception: {}, reply: [{}] {}'.format(
-                        str(e), reply.status_code, reply.text)
-                else:
-                    msg = 'Horizon submit exception: {}'.format(str(e))
-                logging.warning(msg)
+        return check_horizon_reply(reply)
 
-                if (reply is not None and reply.status_code not in self.status_forcelist) or retry_count <= 0:
-                    if reply is None:
-                        raise HorizonRequestError(e)
-                    raise HorizonError('Invalid horizon reply: [{}] {}'.format(
-                        reply.status_code, reply.text), reply.status_code)
-                retry_count -= 1
-                logging.warning('Submit retry attempt {}'.format(retry_count))
-                sleep(self.backoff_factor)
+    async def query(self, rel_url: URL, params: Optional[dict] = None, sse: Optional[bool] = False,
+                    sse_timeout: Optional[Union[float, None]] = None) -> Union[dict, AsyncGenerator]:
+        """
+        Send a query to horizon
+        :param rel_url: The relative path
+        :param params: Parameters to include in the query
+        :param sse: Should SSE be used
+        :param sse_timeout: How long to wait for a new sse event
+        :return: The response from horizon
+        """
+        abs_url = self.horizon_uri.join(rel_url)
+        try:
+            reply = await self._get(abs_url, params, sse, sse_timeout=sse_timeout)
+        except (aiohttp.ClientConnectionError, aiohttp.ContentTypeError) as e:
+                raise HorizonRequestError(e)
 
-    def query(self, rel_url, params=None, sse=False):
-        abs_url = urljoin(self.horizon_uri, rel_url)
-        reply = self._query(abs_url, params, sse)
         return check_horizon_reply(reply) if not sse else reply
 
-    def _query(self, url, params=None, sse=False):
-        reply = None
+    @_retry
+    async def _get(self, url: URL, params: Optional[dict] = None, sse: Optional[bool] = False,
+                    sse_timeout: Optional[Union[float, None]] = None) -> Union[dict, AsyncGenerator]:
+        """
+        Send a get request
+        :param url: The url to send a request to
+        :param params: Parameters to include in the request
+        :param sse: Should SSE be used
+        :param sse_timeout: How long to wait for a new sse event
+        :return: The response from the http request
+        """
         if not sse:
-            try:
-                reply = self._session.get(
-                    url, params=params, timeout=self.request_timeout)
-                return reply.json()
-            except (RequestException, NewConnectionError, ValueError) as e:
-                if reply is not None:
-                    raise HorizonError('Invalid horizon reply: [{}] {}'.format(
-                        reply.status_code, reply.text), reply.status_code)
-                else:
-                    raise HorizonRequestError(e)
+            async with self._session.get(url, params=params) as response:
+                return await response.json(encoding='utf-8')
+        return self.sse_generator(url, sse_timeout)
 
-        # SSE connection
-        if SSEClient is None:
-            raise ImportError('SSE not supported, missing `stellar-base-sseclient` module')
+    @_retry
+    async def _post(self, url: URL, params: Optional[dict] = None) -> dict:
+        """
+        Send a post request
+        :param url: The url to send a request to
+        :param params: Parameters to include
+        :return: The response from the http request
+        """
+        async with self._session.post(url, params=params) as response:
+            return await response.json(encoding='utf-8')
 
-        return SSEClient(url, retry=0, session=self._sse_session, connect_retry=-1, params=params)
+    async def sse_generator(self, url: Union[str, URL], timeout: Union[float, None]) -> AsyncGenerator:
+        """
+        SSE generator with timeout between events
+        :param url: URL to send SSE request to
+        :param timeout: The time to wait for a a new event
+        :return: AsyncGenerator[dict]
+        """
+        async def _sse_generator() -> AsyncGenerator:
+            """
+            Generator for sse events
+            :rtype AsyncGenerator[dict]
+            """
+            last_id = 'now'  # Start monitoring from now.
+            retry = 0.1
+            while True:
+                try:
+                    """
+                    Create a new SSEClient:
+                    Using the last id as the cursor
+                    Headers are needed because of a bug that makes "params" override the default headers
+                    """
+                    async with SSEClient(url, session=self._sse_session,
+                                         params={'cursor': last_id},
+                                         headers=HEADERS) as client:
+                        """
+                        We want to throw a TimeoutError if we didnt get any event in the last x seconds.
+                        read_timeout in aiohttp is not implemented correctly https://github.com/aio-libs/aiohttp/issues/1954
+                        So we will create our own way to do that.
 
-    def account(self, address):
+                        Note that the timeout starts from the first event forward. There is no until we get the first event.
+                        """
+                        async for event in client:
+                            if event.last_event_id != '':
+                                # Events that dont have an id are not useful for us (hello/byebye events)
+                                # Save the last event id and retry time
+                                last_id = event.last_event_id
+                                retry = client._reconnection_time
+                                try:
+                                    yield json.loads(event.data)
+                                except json.JSONDecodeError:
+                                    # Content was not json-decodable
+                                    pass
+                except aiohttp.ClientPayloadError:
+                    # Retry if the connection dropped after we got the initial response
+                    logger.debug('Resetting SSE connection for {} after timeout'.format(url))
+                    await asyncio.sleep(retry.total_seconds())
+
+        await self._init_sse_session()
+        gen = _sse_generator()
+        while True:
+            yield await asyncio.wait_for(gen.__anext__(), timeout)
+
+    async def account(self, address: str) -> dict:
         """Returns information and links relating to a single account.
 
         `GET /accounts/{account}
         <https://www.stellar.org/developers/horizon/reference/endpoints/accounts-single.html>`_
 
-        :param str address: The account ID to retrieve details about.
+        :param address: The account ID to retrieve details about.
         :return: The account details in a JSON response.
         :rtype: dict
 
         """
-        endpoint = '/accounts/{account_id}'.format(account_id=address)
-        return self.query(endpoint)
+        endpoint = URL('/accounts/{account_id}'.format(account_id=address))
+        return await self.query(endpoint)
 
-    def account_data(self, address, key):
+    async def account_data(self, address, key):
         """This endpoint represents a single data associated with a given
         account.
 
@@ -202,11 +259,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/accounts/{account_id}/data/{data_key}'.format(
-            account_id=address, data_key=key)
-        return self.query(endpoint)
+        endpoint = URL('/accounts/{account_id}/data/{data_key}'.format(
+            account_id=address, data_key=key))
+        return await self.query(endpoint)
 
-    def account_effects(self, address, cursor=None, order='asc', limit=10, sse=False):
+    async def account_effects(self, address, cursor=None, order='asc', limit=10, sse=False, sse_timeout=None):
         """This endpoint represents all effects that changed a given account.
 
         `GET /accounts/{account}/effects{?cursor,limit,order}
@@ -223,11 +280,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/accounts/{account_id}/effects'.format(account_id=address)
+        endpoint = URL('/accounts/{account_id}/effects'.format(account_id=address))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def account_offers(self, address, cursor=None, order='asc', limit=10, sse=False):
+    async def account_offers(self, address, cursor=None, order='asc', limit=10, sse=False, sse_timeout=None):
         """This endpoint represents all the offers a particular account makes.
 
         `GET /accounts/{account}/offers{?cursor,limit,order}
@@ -244,11 +301,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/accounts/{account_id}/offers'.format(account_id=address)
+        endpoint = URL('/accounts/{account_id}/offers'.format(account_id=address))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def account_operations(self, address, cursor=None, order='asc', limit=10, sse=False):
+    async def account_operations(self, address, cursor=None, order='asc', limit=10, sse=False, sse_timeout=None):
         """This endpoint represents all operations that were included in valid
         transactions that affected a particular account.
 
@@ -266,57 +323,59 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/accounts/{account_id}/operations'.format(
-            account_id=address)
+        endpoint = URL('/accounts/{account_id}/operations'.format(
+            account_id=address))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def account_transactions(self, address, cursor=None, order='asc', limit=10, sse=False):
+    async def account_transactions(self, address: str, cursor: Optional[int] = None, order: str = 'asc', limit: Optional[int] = 10, sse: Optional[bool] = False,
+                                   sse_timeout: Optional[Union[float, None]] = None) -> Union[dict, AsyncGenerator]:
         """This endpoint represents all transactions that affected a given
         account.
 
         `GET /accounts/{account_id}/transactions{?cursor,limit,order}
         <https://www.stellar.org/developers/horizon/reference/endpoints/transactions-for-account.html>`_
 
-        :param str address: The account ID to list transactions from.
+        :param address: The account ID to list transactions from.
         :param cursor: A paging token, specifying where to start returning records from.
             When streaming this can be set to "now" to stream object created since your request time.
-        :type cursor: int, str
-        :param str order: The order in which to return rows, "asc" or "desc".
-        :param int limit: Maximum number of records to return.
-        :param bool sse: Use server side events for streaming responses.
+        :param order: The order in which to return rows, "asc" or "desc".
+        :param limit: Maximum number of records to return.
+        :param sse: Use server side events for streaming responses.
+        :param sse_timeout: How long to wait between events
         :return: The list of transactions for an account in a JSON response.
         :rtype: dict
 
         """
-        endpoint = '/accounts/{account_id}/transactions'.format(
-            account_id=address)
+        endpoint = URL('/accounts/{account_id}/transactions'.format(
+            account_id=address))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
 
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def account_payments(self, address, cursor=None, order='asc', limit=10, sse=False):
+    async def account_payments(self, address: str, cursor: Optional[int] = None, order: str = 'asc', limit: Optional[int] = 10, sse: Optional[bool] = False,
+                               sse_timeout: Optional[Union[float, None]] = None) -> Union[dict, AsyncGenerator]:
         """This endpoint responds with a collection of Payment operations where
         the given account was either the sender or receiver.
 
         `GET /accounts/{id}/payments{?cursor,limit,order}
         <https://www.stellar.org/developers/horizon/reference/endpoints/payments-for-account.html>`_
 
-        :param str address: The account ID to list payments to/from.
+        :param address: The account ID to list payments to/from.
         :param cursor: A paging token, specifying where to start returning records from.
             When streaming this can be set to "now" to stream object created since your request time.
-        :type cursor: int, str
-        :param str order: The order in which to return rows, "asc" or "desc".
-        :param int limit: Maximum number of records to return.
-        :param bool sse: Use server side events for streaming responses.
+        :param order: The order in which to return rows, "asc" or "desc".
+        :param limit: Maximum number of records to return.
+        :param sse: Use server side events for streaming responses.
+        :param sse_timeout: How long to wait between events
         :return: The list of payments for an account in a JSON response.
         :rtype: dict
         """
-        endpoint = '/accounts/{account_id}/payments'.format(account_id=address)
+        endpoint = URL('/accounts/{account_id}/payments'.format(account_id=address))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def account_trades(self, address, cursor=None, order='asc', limit=10, sse=False):
+    async def account_trades(self, address, cursor=None, order='asc', limit=10, sse=False, sse_timeout=None):
         """This endpoint responds with a collection of Trades where
         the given account was either the taker or the maker
 
@@ -333,11 +392,11 @@ class Horizon(object):
         :return: The list of payments for an account in a JSON response.
         :rtype: dict
         """
-        endpoint = '/accounts/{account_id}/trades'.format(account_id=address)
+        endpoint = URL('/accounts/{account_id}/trades'.format(account_id=address))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def assets(self, asset_code=None, asset_issuer=None, cursor=None, order='asc', limit=10):
+    async def assets(self, asset_code=None, asset_issuer=None, cursor=None, order='asc', limit=10):
         """This endpoint represents all assets. It will give you all the assets
         in the system along with various statistics about each.
 
@@ -358,12 +417,13 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/assets'
+        endpoint = URL('/assets')
         params = self.__query_params(asset_code=asset_code, asset_issuer=asset_issuer, cursor=cursor, order=order,
                                      limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def transactions(self, cursor=None, order='asc', limit=10, sse=False):
+    async def transactions(self, cursor: Optional[int] = None, order: str = 'asc', limit: Optional[int] = 10, sse: Optional[bool] = False,
+                           sse_timeout: Optional[Union[float, None]] = None) -> Union[dict, AsyncGenerator]:
         """This endpoint represents all validated transactions.
 
         `GET /transactions{?cursor,limit,order}
@@ -375,30 +435,31 @@ class Horizon(object):
         :param str order: The order in which to return rows, "asc" or "desc".
         :param int limit: Maximum number of records to return.
         :param bool sse: Use server side events for streaming responses.
+        :param sse_timeout: How long to wait between events
         :return: The list of all transactions
         :rtype: dict
 
         """
-        endpoint = '/transactions'
+        endpoint = URL('/transactions')
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def transaction(self, tx_hash):
+    async def transaction(self, tx_hash: str) -> dict:
         """The transaction details endpoint provides information on a single
         transaction.
 
         `GET /transactions/{hash}
         <https://www.stellar.org/developers/horizon/reference/endpoints/transactions-single.html>`_
 
-        :param str tx_hash: The hex-encoded transaction hash.
+        :param tx_hash: The hex-encoded transaction hash.
         :return: A single transaction's details.
         :rtype: dict
 
         """
-        endpoint = '/transactions/{tx_hash}'.format(tx_hash=tx_hash)
-        return self.query(endpoint)
+        endpoint = URL('/transactions/{tx_hash}'.format(tx_hash=tx_hash))
+        return await self.query(endpoint)
 
-    def transaction_operations(self, tx_hash, cursor=None, order='asc', limit=10):
+    async def transaction_operations(self, tx_hash, cursor=None, order='asc', limit=10):
         """This endpoint represents all operations that are part of a given
         transaction.
 
@@ -413,11 +474,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/transactions/{tx_hash}/operations'.format(tx_hash=tx_hash)
+        endpoint = URL('/transactions/{tx_hash}/operations'.format(tx_hash=tx_hash))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def transaction_effects(self, tx_hash, cursor=None, order='asc', limit=10):
+    async def transaction_effects(self, tx_hash, cursor=None, order='asc', limit=10):
         """This endpoint represents all effects that occurred as a result of a
         given transaction.
 
@@ -432,11 +493,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/transactions/{tx_hash}/effects'.format(tx_hash=tx_hash)
+        endpoint = URL('/transactions/{tx_hash}/effects'.format(tx_hash=tx_hash))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def transaction_payments(self, tx_hash, cursor=None, order='asc', limit=10):
+    async def transaction_payments(self, tx_hash, cursor=None, order='asc', limit=10):
         """This endpoint represents all payment operations that are part of a
         given transaction.
 
@@ -451,11 +512,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/transactions/{tx_hash}/payments'.format(tx_hash=tx_hash)
+        endpoint = URL('/transactions/{tx_hash}/payments'.format(tx_hash=tx_hash))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def order_book(self, selling_asset_code, buying_asset_code, selling_asset_issuer=None, buying_asset_issuer=None,
+    async def order_book(self, selling_asset_code, buying_asset_code, selling_asset_issuer=None, buying_asset_issuer=None,
                    limit=10):
         """Return, for each orderbook, a summary of the orderbook and the bids
         and asks associated with that orderbook.
@@ -486,11 +547,11 @@ class Horizon(object):
             'buying_asset_code': None if buying_asset.is_native() else buying_asset.code,
             'buying_asset_issuer': buying_asset.issuer,
         }
-        endpoint = '/order_book'
+        endpoint = URL('/order_book')
         params = self.__query_params(limit=limit, **asset_params)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def ledgers(self, cursor=None, order='asc', limit=10, sse=False):
+    async def ledgers(self, cursor=None, order='asc', limit=10, sse=False, sse_timeout=None):
         """This endpoint represents all ledgers.
 
         `GET /ledgers{?cursor,limit,order}
@@ -506,25 +567,25 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/ledgers'
+        endpoint = URL('/ledgers')
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def ledger(self, ledger_id):
+    async def ledger(self, ledger_id: int) -> dict:
         """The ledger details endpoint provides information on a single ledger.
 
         `GET /ledgers/{sequence}
         <https://www.stellar.org/developers/horizon/reference/endpoints/ledgers-single.html>`_
 
-        :param int ledger_id: The id of the ledger to look up.
+        :param ledger_id: The id of the ledger to look up.
         :return: The details of a single ledger.
         :rtype: dict
 
         """
-        endpoint = '/ledgers/{ledger_id}'.format(ledger_id=ledger_id)
-        return self.query(endpoint)
+        endpoint = URL('/ledgers/{ledger_id}'.format(ledger_id=ledger_id))
+        return await self.query(endpoint)
 
-    def ledger_transactions(self, ledger_id, cursor=None, order='asc', limit=10):
+    async def ledger_transactions(self, ledger_id, cursor=None, order='asc', limit=10):
         """This endpoint represents all transactions that occurred in the given
         ledger.
 
@@ -539,11 +600,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/ledgers/{ledger_id}/transactions'.format(ledger_id=ledger_id)
+        endpoint = URL('/ledgers/{ledger_id}/transactions'.format(ledger_id=ledger_id))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def ledger_effects(self, ledger_id, cursor=None, order='asc', limit=10):
+    async def ledger_effects(self, ledger_id, cursor=None, order='asc', limit=10):
         """This endpoint represents all effects that occurred in the given
         ledger.
 
@@ -558,11 +619,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/ledgers/{ledger_id}/effects'.format(ledger_id=ledger_id)
+        endpoint = URL('/ledgers/{ledger_id}/effects'.format(ledger_id=ledger_id))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def ledger_operations(self, ledger_id, cursor=None, order='asc', limit=10):
+    async def ledger_operations(self, ledger_id, cursor=None, order='asc', limit=10):
         """This endpoint returns all operations that occurred in a given
         ledger.
 
@@ -577,12 +638,12 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/ledgers/{ledger_id}/operations'.format(
-            ledger_id=ledger_id)
+        endpoint = URL('/ledgers/{ledger_id}/operations'.format(
+            ledger_id=ledger_id))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def ledger_payments(self, ledger_id, cursor=None, order='asc', limit=10):
+    async def ledger_payments(self, ledger_id, cursor=None, order='asc', limit=10):
         """This endpoint represents all payment operations that are part of a
         valid transactions in a given ledger.
 
@@ -597,30 +658,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/ledgers/{ledger_id}/payments'.format(ledger_id=ledger_id)
+        endpoint = URL('/ledgers/{ledger_id}/payments'.format(ledger_id=ledger_id))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def ledger_transactions(self, ledger_id, cursor=None, order='asc', limit=10):
-        """This endpoint represents all transactions in a given ledger.
-
-        `GET /ledgers/{id}/transactions{?cursor,limit,order}
-        <https://www.stellar.org/developers/horizon/reference/endpoints/transactions-for-ledger.html>`_
-
-        :param int ledger_id: The id of the ledger to look up.
-        :param int cursor: A paging token, specifying where to start returning records from.
-        :param str order: The order in which to return rows, "asc" or "desc".
-        :param int limit: Maximum number of records to return.
-        :return: The transactions contained in a single ledger.
-        :rtype: dict
-
-        """
-        endpoint = '/ledgers/{ledger_id}/transactions'.format(
-            ledger_id=ledger_id)
-        params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
-
-    def effects(self, cursor=None, order='asc', limit=10, sse=False):
+    async def effects(self, cursor=None, order='asc', limit=10, sse=False, sse_timeout=None):
         """This endpoint represents all effects.
 
         `GET /effects{?cursor,limit,order}
@@ -636,11 +678,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/effects'
+        endpoint = URL('/effects')
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def operations(self, cursor=None, order='asc', limit=10, sse=False):
+    async def operations(self, cursor=None, order='asc', limit=10, sse=False, sse_timeout=None):
         """This endpoint represents all operations that are part of validated
         transactions.
 
@@ -657,11 +699,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/operations'
+        endpoint = URL('/operations')
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def operation(self, op_id):
+    async def operation(self, op_id):
         """The operation details endpoint provides information on a single
         operation.
 
@@ -672,10 +714,10 @@ class Horizon(object):
         :return: Details on a single operation.
         :rtype: dict
         """
-        endpoint = '/operations/{op_id}'.format(op_id=op_id)
-        return self.query(endpoint)
+        endpoint = URL('/operations/{op_id}'.format(op_id=op_id))
+        return await self.query(endpoint)
 
-    def operation_effects(self, op_id, cursor=None, order='asc', limit=10):
+    async def operation_effects(self, op_id, cursor=None, order='asc', limit=10):
         """This endpoint represents all effects that occurred as a result of a
         given operation.
 
@@ -690,11 +732,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/operations/{op_id}/effects'.format(op_id=op_id)
+        endpoint = URL('/operations/{op_id}/effects'.format(op_id=op_id))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def payments(self, cursor=None, order='asc', limit=10, sse=False):
+    async def payments(self, cursor=None, order='asc', limit=10, sse=False, sse_timeout=None):
         """This endpoint represents all payment operations that are part of
         validated transactions.
 
@@ -711,11 +753,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/payments'
+        endpoint = URL('/payments')
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params, sse)
+        return await self.query(endpoint, params, sse, sse_timeout)
 
-    def paths(self, destination_account, destination_amount, source_account, destination_asset_code,
+    async def paths(self, destination_account, destination_amount, source_account, destination_asset_code,
               destination_asset_issuer=None):
         """Load a list of assets available to the source account id and find
         any payment paths from those source assets to the desired
@@ -747,15 +789,15 @@ class Horizon(object):
             'destination_asset_code': None if destination_asset.is_native() else destination_asset.code,
             'destination_asset_issuer': destination_asset.issuer
         }
-        endpoint = '/paths'
+        endpoint = URL('/paths')
         params = self.__query_params(destination_account=destination_account,
                                      source_account=source_account,
                                      destination_amount=destination_amount,
                                      **destination_asset_params
                                      )
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def trades(self, base_asset_code=None, counter_asset_code=None, base_asset_issuer=None, counter_asset_issuer=None,
+    async def trades(self, base_asset_code=None, counter_asset_code=None, base_asset_issuer=None, counter_asset_issuer=None,
                offer_id=None, cursor=None, order='asc', limit=10):
         """Load a list of trades, optionally filtered by an orderbook.
 
@@ -787,11 +829,11 @@ class Horizon(object):
             'counter_asset_code': None if counter_asset.is_native() else counter_asset.code,
             'counter_asset_issuer': counter_asset.issuer
         }
-        endpoint = '/trades'
+        endpoint = URL('/trades')
         params = self.__query_params(offer_id=offer_id, cursor=cursor, order=order, limit=limit, **asset_params)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def trade_aggregations(self, resolution, base_asset_code, counter_asset_code,
+    async def trade_aggregations(self, resolution, base_asset_code, counter_asset_code,
                            base_asset_issuer=None, start_time=None, end_time=None,
                            counter_asset_issuer=None, order='asc', limit=10):
         """Load a list of aggregated historical trade data, optionally filtered
@@ -825,12 +867,12 @@ class Horizon(object):
             'counter_asset_code': None if counter_asset.is_native() else counter_asset.code,
             'counter_asset_issuer': counter_asset.issuer
         }
-        endpoint = '/trade_aggregations'
+        endpoint = URL('/trade_aggregations')
         params = self.__query_params(start_time=start_time, end_time=end_time, resolution=resolution, order=order,
                                      limit=limit, **asset_params)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def offer_trades(self, offer_id, cursor=None, order='asc', limit=10):
+    async def offer_trades(self, offer_id, cursor=None, order='asc', limit=10):
         """This endpoint represents all trades for a given offer.
 
         `GET /offers/{offer_id}/trades{?cursor,limit,order}
@@ -844,11 +886,11 @@ class Horizon(object):
         :rtype: dict
 
         """
-        endpoint = '/offers/{offer_id}/trades'.format(offer_id=offer_id)
+        endpoint = URL('/offers/{offer_id}/trades'.format(offer_id=offer_id))
         params = self.__query_params(cursor=cursor, order=order, limit=limit)
-        return self.query(endpoint, params)
+        return await self.query(endpoint, params)
 
-    def metrics(self):
+    async def metrics(self):
         """The metrics endpoint returns a host of useful data points for monitoring the health
         of the underlying Horizon process.
 
@@ -859,25 +901,32 @@ class Horizon(object):
         :rtype: dict
         """
 
-        endpoint = '/metrics'
-        return self.query(endpoint)
+        endpoint = URL('/metrics')
+        return await self.query(endpoint)
 
-    def __query_params(self, **kwargs):
+    @staticmethod
+    def __query_params(**kwargs) -> Union[dict, None]:
         params = {k: v for k, v in kwargs.items() if v is not None}
         return params
 
+    async def close(self) -> None:
+        """Close the connection to horizon"""
+        await self._session.__aexit__(None, None, None)
+        if self._sse_session is not None:
+            await self._sse_session.__aexit__(None, None, None)
 
-def check_horizon_reply(reply):
+
+def check_horizon_reply(reply: dict) -> dict:
     if 'status' not in reply:
         return reply
     raise HorizonError(reply)
 
 
-def horizon_testnet():
-    """Create a Horizon instance utilizing SDF's Test Network."""
+async def horizon_testnet() -> 'Horizon':
+    """Create a Horizon instance utilizing Kin's Test Network."""
     return Horizon(HORIZON_TEST)
 
 
-def horizon_livenet():
-    """Create a Horizon instance utilizing SDF's Live Network."""
+async def horizon_livenet() -> 'Horizon':
+    """Create a Horizon instance utilizing Kin's Live Network."""
     return Horizon(HORIZON_LIVE)
