@@ -1,15 +1,15 @@
-import json
+import asyncio
+
 import pytest
-import requests
+from aiohttp import ClientSession
 
 from kin_base.asset import Asset
 from kin_base.keypair import Keypair
 from kin_base.builder import Builder
+from kin_base.horizon import Horizon
 
 import logging
 logging.basicConfig()
-
-#logging.getLogger().setLevel(logging.DEBUG)
 
 
 def pytest_addoption(parser):
@@ -49,16 +49,23 @@ def setup():
         friendbot_url='http://localhost:8001')
 
 
+@pytest.yield_fixture(scope='session')
+async def aio_session():
+    session = ClientSession()
+    yield session
+    await session.close()
+
+
 class Helpers:
     """A container for helper functions available to all tests"""
 
     @staticmethod
-    def fund_account(setup, address):
+    async def fund_account(setup, address, aio_session):
         for attempt in range(3):
             try:
-                r = requests.get(setup.friendbot_url +
-                                 '?addr=' + address)
-                j = json.loads(r.text)
+                async with aio_session.get(setup.friendbot_url +
+                                 '?addr=' + address) as r:
+                    j = await r.json()
                 if 'hash' in j:
                     print('\naccount {} funded successfully'.format(address))
                     return
@@ -69,46 +76,57 @@ class Helpers:
                 else:
                     raise Exception('unexpected friendbot reply')
             except Exception as e:
+                print(e)
                 print('\naccount {} funding error: {} {}'.format(
-                    address, r.status_code, r.text))
+                    address, r.status, await r.text()))
         raise Exception('account {} funding failed'.format(address))
 
     @staticmethod
-    def trust_asset(setup, secret_key, memo_text=None):
+    async def trust_asset(setup, secret_key, memo_text=None):
         """A helper to establish a trustline"""
-        builder = Builder(
-            secret=secret_key,
-            horizon_uri=setup.horizon_endpoint_uri,
-            network=setup.network)
-        builder.append_trust_op(setup.test_asset.issuer, setup.test_asset.code)
-        if memo_text:
-            builder.add_text_memo(memo_text[:28])  # max memo length is 28
-        builder.sign()
-        reply = builder.submit()
+        async with Horizon(setup.horizon_endpoint_uri) as horizon:
+            builder = Builder(
+                secret=secret_key,
+                horizon=horizon,
+                network_name=setup.network,
+                fee=100)
+            builder.append_trust_op(setup.test_asset.issuer, setup.test_asset.code)
+            if memo_text:
+                builder.add_text_memo(memo_text[:28])  # max memo length is 28
+            builder.sign()
+            reply = await builder.submit()
         return reply.get('hash')
 
     @classmethod
-    def fund_asset(cls, setup, address, amount, memo_text=None):
+    async def fund_asset(cls, setup, address, amount, memo_text=None):
         """A helper to fund account with test asset"""
         return cls.send_asset(setup, setup.issuer_keypair.seed(), address,
                               amount, memo_text)
 
     @classmethod
-    def send_asset(cls, setup, secret_key, address, amount, memo_text=None):
+    async def send_asset(cls, setup, secret_key, address, amount, memo_text=None):
         """A helper to send asset"""
-        builder = Builder(
-            secret=secret_key,
-            horizon_uri=setup.horizon_endpoint_uri,
-            network=setup.network)
-        builder.append_payment_op(address, amount, setup.test_asset.code,
-                                  setup.test_asset.issuer)
-        if memo_text:
-            builder.add_text_memo(memo_text[:28])  # max memo length is 28
-        builder.sign()
-        reply = builder.submit()
+        async with Horizon(setup.horizon_endpoint_uri) as horizon:
+            builder = Builder(
+                secret=secret_key,
+                horizon=horizon,
+                network_name=setup.network,
+                fee=100)
+            builder.append_payment_op(address, amount, setup.test_asset.code,
+                                      setup.test_asset.issuer)
+            if memo_text:
+                builder.add_text_memo(memo_text[:28])  # max memo length is 28
+            builder.sign()
+            reply = await builder.submit()
         return reply.get('hash')
 
 
 @pytest.fixture(scope='session')
 def helpers():
     return Helpers
+
+@pytest.yield_fixture(scope='session')
+def event_loop():
+    loop = asyncio.get_event_loop_policy().new_event_loop()
+    yield loop
+    loop.close()
