@@ -12,8 +12,8 @@ from kin_base.horizon import Horizon
 # TODO: These endpoints really need to be mocked out.
 
 
-@pytest.fixture(scope='module')
-def test_data(setup, helpers):
+@pytest.yield_fixture(scope='module')
+async def test_data(setup, helpers, aio_session):
     class Struct:
         def __init__(self, **entries):
             self.__dict__.update(entries)
@@ -21,13 +21,16 @@ def test_data(setup, helpers):
     cold = Keypair.random()
     cold_secret = cold.seed()
     cold_account = cold.address().decode()
+    horizon = Horizon(setup.horizon_endpoint_uri)
 
-    helpers.fund_account(setup, cold.address().decode())
+    await helpers.fund_account(setup, cold.address().decode(), aio_session)
 
-    return Struct(cold_secret=cold_secret, cold_account=cold_account, horizon=Horizon(setup.horizon_endpoint_uri))
+    yield Struct(cold_secret=cold_secret, cold_account=cold_account, horizon=horizon)
+    await horizon.close()
 
 
-def test_builder_(setup, test_data):
+@pytest.mark.asyncio
+async def test_builder_(setup, test_data):
     hot = Keypair.random()
     hot_account = hot.address().decode()
     hot_secret = hot.seed()
@@ -43,7 +46,7 @@ def test_builder_(setup, test_data):
     cold.append_payment_op(hot_account, '50.123', 'BEER', test_data.cold_account) \
         .append_payment_op(hot_account, '50.123', 'BEER', test_data.cold_account)
     # TODO: append_bump_sequence_op test
-    cold.update_sequence()
+    await cold.update_sequence()
     cold.sign()
     cold.sign(hot_secret)
     # try to sign twice
@@ -53,23 +56,24 @@ def test_builder_(setup, test_data):
     assert len(cold.te.signatures) == 2
     assert len(cold.ops) == 5
 
-    response = cold.submit()
+    response = await cold.submit()
     assert response.get('hash') == cold.hash_hex()
 
 
-def test_builder_xdr(setup, helpers, test_data):
+@pytest.mark.asyncio
+async def test_builder_xdr(setup, helpers, test_data, aio_session):
     hot = Keypair.random()
     hot_account = hot.address().decode()
     hot_secret = hot.seed()
 
-    helpers.fund_account(setup, hot_account)
+    await helpers.fund_account(setup, hot_account, aio_session)
 
     cold = Builder(secret=test_data.cold_secret, horizon=test_data.horizon, network_name=setup.network, fee=100) \
         .append_change_trust_op('BEER', test_data.cold_account, '1000', hot_account) \
         .append_allow_trust_op(hot_account, 'BEER', True, test_data.cold_account) \
         .append_payment_op(hot_account, '100', 'BEER', test_data.cold_account, test_data.cold_account) \
         .append_payment_op(test_data.cold_account, '2.222', 'BEER', test_data.cold_account, hot_account)
-    cold.update_sequence()
+    await cold.update_sequence()
     cold.sign()
 
     xdr = cold.gen_xdr()
@@ -84,7 +88,7 @@ def test_builder_xdr(setup, helpers, test_data):
     assert len(hot.te.signatures) == 2
     assert len(hot.ops) == 4
 
-    response = hot.submit()
+    response = await hot.submit()
     assert response.get('hash') == hot.hash_hex()
 
 
@@ -113,28 +117,31 @@ def test_builder(test_data, setup):
     return builder
 
 
-def test_sign(test_builder):
+@pytest.mark.asyncio
+async def test_sign(test_builder):
     address = 'GCAZ7QXD6UJ5NOVWYTNKLNP36DPJZMRO67LQ4X5CH2IHY3OG5QGECGYQ'
 
     test_builder.append_create_account_op(address, '100')
     assert len(test_builder.ops) == 1
-    test_builder.update_sequence()
+    await test_builder.update_sequence()
     test_builder.sign()
     assert test_builder.te
     assert test_builder.tx
 
 
-def test_update_sequence(test_builder):
-    test_builder.update_sequence()
+@pytest.mark.asyncio
+async def test_update_sequence(test_builder):
+    await test_builder.update_sequence()
     assert test_builder.sequence
 
 
-def test_set_channel(setup, helpers, test_builder):
+@pytest.mark.asyncio
+async def test_set_channel(setup, helpers, test_builder, aio_session):
     channel_addr = 'GBC6PXY4ZSO356NUPF2A2SDVEBQB2RG7XN6337NBW4F24APGHEVR3IIU'
     channel = 'SA4BHY26Q3C3BSYKKGDM7UVMZ4YF6YBLX6AOWYEDBXPLOR7WQ5EJXN6X'
-    helpers.fund_account(setup, channel_addr)
+    await helpers.fund_account(setup, channel_addr, aio_session)
     time.sleep(5)
-    test_builder.set_channel(channel)
+    await test_builder.set_channel(channel)
     assert test_builder.address == channel_addr
     assert test_builder.sequence
 
